Improve step definition to fix testing issue
The problem here was that this step was returning straight away, and
then Frank was moving on to the next testing step while the animation
from one screen to the other was still 'in flight'. That was causing two
issues:

1) The 'navigate backwards' step wasn't working, probably because the
OS isn't expecting someone to click on the back button as the main view
is still animating in.

2) The 'I should see "Edit"' step was passing even though it shouldn't.
That's probably because the Edit button was still in the view heirarchy
as the animation was still in flight.